# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.bulkIDsRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
...
  controllers.createMessageRequest:
    properties:
      content:
        type: string
      roleIds:
        items:
          type: integer
        type: array
      targetType:
        enum:
        - ALL
        - USER
        - ROLE
        type: string
      title:
        type: string
      type:
        enum:
        - SYSTEM
        - NOTICE
        - ANNOUNCEMENT
        type: string
      userIds:
        items:
          type: integer
        type: array
    required:
    - content
    - targetType
    - title
    - type
    type: object
...
  models.Message:
    properties:
      content:
        type: string
      createTime:
        type: string
      id:
        type: integer
      senderId:
        type: integer
      senderName:
        type: string
      status:
        type: string
      targetRoles:
        description: JSON array of role IDs, e.g. "[1,2]"
        type: string
      targetType:
        type: string
      title:
        type: string
      type:
        type: string
      updateTime:
        type: string
    type: object
  models.PageData:
    properties:
      pageData: {}
      total:
        type: integer
    type: object
...
  models.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      originUrl:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: kite-admin 后台管理系统 API 文档
  title: Kite Admin API
  version: "1.0"
paths:
...
  /message:
    post:
      consumes:
      - application/json
      description: 创建并发送消息，支持按用户、角色或全员发送
      parameters:
      - description: 消息信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.createMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Message'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 发送消息
      tags:
      - 消息管理
  /message/{id}:
    delete:
      description: 删除指定消息及其收件人记录
      parameters:
      - description: 消息ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 删除消息
      tags:
      - 消息管理
  /message/{id}/read:
    patch:
      description: 标记指定消息为已读
      parameters:
      - description: 消息ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: 消息不存在
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 标记消息已读
      tags:
      - 消息管理
  /message/bulk/delete:
    post:
      consumes:
      - application/json
      description: 批量删除指定消息
      parameters:
      - description: 消息ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.bulkIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除数量
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  type: integer
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 批量删除消息
      tags:
      - 消息管理
  /message/bulk/read:
    patch:
      consumes:
      - application/json
      description: 批量标记指定消息为已读
      parameters:
      - description: 消息ID列表
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.bulkIDsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 批量标记已读
      tags:
      - 消息管理
  /message/mine:
    get:
      description: 获取当前用户收到的消息列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: pageNo
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 获取我的消息
      tags:
      - 消息管理
  /message/page:
    get:
      description: 分页查询消息列表，支持按标题和类型筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: pageNo
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 标题（模糊搜索）
        in: query
        name: title
        type: string
      - description: 类型（SYSTEM/NOTICE/ANNOUNCEMENT）
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  allOf:
                  - $ref: '#/definitions/models.PageData'
                  - properties:
                      pageData:
                        items:
                          $ref: '#/definitions/models.Message'
                        type: array
                    type: object
              type: object
      security:
      - BearerAuth: []
      summary: 分页查询消息
      tags:
      - 消息管理
  /message/read/all:
    patch:
      description: 将当前用户所有未读消息标记为已读
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 全部标记已读
      tags:
      - 消息管理
  /message/sse:
    get:
      description: 建立SSE连接，实时接收新消息推送
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE事件流
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 消息推送SSE连接
      tags:
      - 消息管理
  /message/unread/count:
    get:
      description: 获取当前用户的未读消息数量
      produces:
      - application/json
      responses:
        "200":
          description: 成功，data包含count字段
          schema:
            $ref: '#/definitions/models.Response'
      security:
      - BearerAuth: []
      summary: 获取未读消息数
      tags:
      - 消息管理
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"